posts/creating-a-personal-blog-part-3-adding-a-contact-form-in-django/source code/Blog/requirements.txt
django==2.1.7
django-simple-captcha==0.5.11